Refactoring to not require the type argument, calling useIsReceiptPage() instead. Moving things around to make tests/mocking work.
