+++ src/core/routing/types.ts
@@ -1,0 +1,7 @@
+export enum SearchParams {
+  FocusComponentId = 'focusComponentId',
+  FocusErrorBinding = 'focusErrorBinding',
+  ExitSubform = 'exitSubform',
+  Validate = 'validate',
+  Pdf = 'pdf',
+}
