+++ src/layout/RepeatingGroup/EditContainer/RepeatingGroupEditContext.tsx
@@ -3,8 +3,8 @@ import { useSearchParams } from 'react-router-dom';
 import type { PropsWithChildren } from 'react';
 
 import { createContext } from 'src/core/contexts/context';
+import { SearchParams } from 'src/core/routing/types';
 import { useLayoutLookups } from 'src/features/form/layout/LayoutsContext';
-import { SearchParams } from 'src/hooks/navigation';
 import { useMemoDeepEqual } from 'src/hooks/useStateDeepEqual';
 import { useRepeatingGroupComponentId } from 'src/layout/RepeatingGroup/Providers/RepeatingGroupContext';
 import { RepGroupHooks } from 'src/layout/RepeatingGroup/utils';
