+++ src/layout/RepeatingGroup/Providers/RepeatingGroupFocusContext.tsx
@@ -3,10 +3,10 @@ import { useSearchParams } from 'react-router-dom';
 import type { PropsWithChildren } from 'react';
 
 import { createContext } from 'src/core/contexts/context';
+import { SearchParams } from 'src/core/routing/types';
 import { useLayoutLookups } from 'src/features/form/layout/LayoutsContext';
 import { isRepeatingComponentType } from 'src/features/form/layout/utils/repeating';
 import { FD } from 'src/features/formData/FormDataWrite';
-import { SearchParams } from 'src/hooks/navigation';
 import {
   RepGroupContext,
   useRepeatingGroupComponentId,
