+++ src/layout/Subform/SubformWrapper.tsx
@@ -9,7 +9,6 @@ import { FormProvider } from 'src/features/form/FormContext';
 import { useDataTypeFromLayoutSet } from 'src/features/form/layout/LayoutsContext';
 import { useNavigationParam } from 'src/hooks/navigation';
 import { useNavigatePage } from 'src/hooks/useNavigatePage';
-import { ProcessTaskType } from 'src/types';
 import { useItemWhenType } from 'src/utils/layout/useNodeItem';
 
 export function SubformWrapper({ baseComponentId, children }: PropsWithChildren<{ baseComponentId: string }>) {
@@ -24,7 +23,7 @@ export function SubformWrapper({ baseComponentId, children }: PropsWithChildren<
 
 export function SubformForm() {
   return (
-    <PresentationComponent type={ProcessTaskType.Data}>
+    <PresentationComponent>
       <Form />
     </PresentationComponent>
   );
